+++ sc/tests/productdownload.py
@@ -1,0 +1,28 @@
+import unittest
+
+from sc.app.productdownload import ProductGrabber, geo_json
+
+class TestProductGrabber(unittest.TestCase):
+    '''
+    Test functions for the ProductGrabber class
+    '''
+    def test_initProductGrabber(self):
+        '''
+        Test product grabber initialization
+        Fails when there is an error in the code
+        '''
+        pg = ProductGrabber()
+        
+    def test_getJSONFeed(self):
+        '''
+        Tests access to the USGS JSON feed. Failure points to error in
+        code, lack of internet access, or a down USGS server
+        '''
+        pg = ProductGrabber()
+        pg.get_json_feed()
+        
+        self.assertNotEqual(pg.json_feed, '')
+
+    def test_geoJSON(self):
+        result = geo_json('hour')
+        self.assertEqual(result['error'], '')
